+++ examples/standard-tf/output.tf
@@ -1,0 +1,28 @@
+
+output "url" {
+  value = module.wandb.url
+}
+
+output "address" {
+  value = module.wandb.address
+}
+
+output "bucket_name" {
+  value = module.wandb.bucket_name
+}
+
+output "standardized_size" {
+  value = var.size
+}
+
+output "gke_node_count" {
+  value = module.wandb.gke_node_count
+}
+
+output "gke_node_instance_type" {
+  value = module.wandb.gke_node_instance_type
+}
+
+output "database_instance_type" {
+  value = module.wandb.database_instance_type
+}