+++ modules/app_gke/variables.tf
@@ -22,6 +22,30 @@ variable "machine_type" {
   type = string
 }
 
+variable "weave_wandb_env" {
+  type        = map(string)
+  description = "Extra environment variables for W&B"
+  default     = {}
+}
+
+variable "app_wandb_env" {
+  type        = map(string)
+  description = "Extra environment variables for W&B"
+  default     = {}
+}
+
+variable "parquet_wandb_env" {
+  type        = map(string)
+  description = "Extra environment variables for W&B"
+  default     = {}
+}
+
+
 variable "node_count" {
   type = number
+}
+
+variable "create_workload_identity" {
+  description = "Flag to indicate whether to enable workload identity for the service account."
+  type        = bool
 }