+++ outputs.tf
@@ -77,6 +77,7 @@ output "gke_node_count" {
   value = coalesce(try(local.deployment_size[var.size].node_count, null), var.gke_node_count)
 }
 
+
 output "gke_node_instance_type" {
   value = coalesce(try(local.deployment_size[var.size].node_instance, null), var.gke_machine_type)
 }
@@ -85,5 +86,7 @@ output "database_instance_type" {
   value = coalesce(try(local.deployment_size[var.size].db, null), var.database_machine_type)
 }
 
-
-
+output "sa_account_email" {
+  description =  "This output provides the email address of the service account created for workload identity, if workload identity is enabled. Otherwise, it returns null"
+  value = var.create_workload_identity == true ? module.service_accounts.sa_account_email : null
+}